refactor: refactor demo main script file to use es6
The older syntax was used because the script was initially
inlined so it was necessary to support older browsers (IE). The script
grew in size however and it made sense to break it into a separate file.
So now we can use es6 and let babel take care of the transpilation.